feat(whatsapp): implements event handler service
This commit introduces a new service, `WhatsappEventService`, which allows for binding custom event handlers to the WhatsApp connection.

The service uses NestJS's `DiscoveryService` to identify methods that are annotated with the `@WAEvent` decorator and bind them to the appropriate event. This allows for a more modular and extensible way of handling WhatsApp events.

Additionally, a new decorator, `@WAEvent`, has been created to simplify the process of annotating methods as event handlers. This decorator takes the event name as a parameter and adds the `WhatsappEventActionMetadataKey` metadata to the method.