feat(whatsapp): add simple menu action
This commit adds a new "Simple Menu" action to the WhatsApp bot. This action allows users to access a menu of all available actions by sending a message with the "menu" flag.

The menu action is implemented using the new `SimpleMenuAction` class, which extends the `WhatsappMessageAction` class. The `SimpleMenuAction` class defines the `execute` method, which is called when the action is triggered. The `execute` method sends a message to the user containing a list of all available actions and their corresponding flags.

This action can be used to improve the user experience of the WhatsApp bot by making it easier for users to discover and use all of the available actions.