chore: add jne resi checking feature
Add a new feature to check JNE resi status. This feature is added in the whatsapp-action module.

The JNE resi checking feature can be used by sending a message with the format "jne <resi>". The resi number can be obtained from the JNE website or from the JNE resi notification.

The JNE resi checking feature will return the following information:

- Resi number
- Delivery status
- Estimated delivery date
- Proof of delivery (POD) date
- Link to the JNE website for more information